refactor db module
