+++ tracker/tot_balances.py
@@ -1,21 +1,16 @@
-import copy
-from datetime import datetime, timedelta
-from lib2to3.pgen2.tokenize import TokenError
 from typing import List
 
 from tracker import utils
-from tracker.db_handler import *
+from tracker import db
 from tracker.telegram_handler import send_message
 
-# from tracker.coingecko_api import *
-import re
 import operator
 import itertools
 
 
 def tot_balances_get():
 
-    missing_tot_bals = db_get_missing_tot_balances()
+    missing_tot_bals = db.balance.db_get_missing_tot_balances()
 
     # group missing total balances by date
     grouped_missing = []
@@ -45,13 +40,15 @@ def tot_balances_get():
 
 def tot_balances_save_on_db(tot_balances: List[dict]):
     for b in tot_balances:
-        db_add_tot_balance(b["date"], b["eur_amount"], b["usd_amount"])
+        db.balance.db_add_tot_balance(b["date"], b["eur_amount"], b["usd_amount"])
 
 
 def get_day_details(date: str) -> str:
     # details
-    num_coins = {c["coin_id"]: c["amount"] for c in db_get_coin_balances(date)}
-    val_coins = {c["coin_id"]: c["coin_eur"] for c in db_get_prices(date)}
+    num_coins = {
+        c["coin_id"]: c["amount"] for c in db.balance.db_get_coin_balances(date)
+    }
+    val_coins = {c["coin_id"]: c["coin_eur"] for c in db.price.db_get_prices(date)}
 
     tot_eur = 0
     last_day_amts = []
@@ -64,7 +61,7 @@ def get_day_details(date: str) -> str:
 
     for c in last_day_amts:
         c["percentage"] = c["eur_amount"] / tot_eur
-        c["coin_name"] = db_get_coin_symbol_by_id(c["coin_id"])
+        c["coin_name"] = db.coin.db_get_coin_symbol_by_id(c["coin_id"])
 
     # order by relevance
     last_day_amts = sorted(last_day_amts, key=lambda d: d["percentage"], reverse=True)
